Refactor project folder
