+++ client/package.json
@@ -14,7 +14,6 @@
     "@emotion/styled": "^11.11.0",
     "@hookform/resolvers": "^3.1.1",
     "@mui/icons-material": "^5.11.16",
-   
     "@mui/material": "^5.13.0",
     "@mui/styled-engine-sc": "^5.12.0",
     "@mui/system": "^5.13.2",
@@ -27,6 +26,7 @@
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-bootstrap-daterangepicker": "^8.0.0",
+    "react-chatbot-kit": "^2.1.2",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.45.0",
     "react-icons": "^4.8.0",
