+++ client/src/components/chat/Chatbot.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Chatbot = () => {
+  return (
+    <div>Chatbot</div>
+  )
+}
+
+export default Chatbot
