+++ client/src/components/pages/budget/Budget.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import Section from "../../layout/Section/Section";
-
+import Chatbot from "../../chat/Chatbot";
 const Budget = () => {
   return (
     <Section title="Budget">
-       
+       <Chatbot/>
     </Section>
   );
 };
