+++ client/src/components/pages/wallet/WalletsPage.jsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from 'react';
 //components
 import Stack from '@mui/material/Stack';
 import CustomTabs from '../../general/CustomTabs';
-import TableData from '../../layout/TableData/TableData';
-import WalletContent from '../WalletsView/WalletContent';
+import TableData from '../../general/TableData/TableData';
+import WalletContent from './components/WalletsView/WalletContent';
 import Section from '../../layout/Section/Section';
 
 //api
