--- package.json
@@ -1,10 +1,0 @@
-{
-  "dependencies": {
-    "axios": "^1.4.0",
-    "firebase": "^9.22.2",
-    "uuid": "^9.0.0"
-  },
-  "devDependencies": {
-    "@faker-js/faker": "^8.0.2"
-  }
-}
