+++ server/src/router/income.js
@@ -1,35 +1,8 @@
-const {PrismaClient} = require('@prisma/client')
 const express = require('express')
-
-const prisma = new PrismaClient()
 const router = express.Router()
 
-router.get('/total', async (req, res) => {
-    
-    try {
-       
-        const incomes = await prisma.transaction.findMany({
-            where: {
-                type: "income"
-            }
-        })
-        
-      
-        // sum all the incomes
-        let totalIncome = 0
-        
-        for (let i = 0; i < incomes.length; i++) {
-            totalIncome += incomes[i].amount
-        }
+const {getIncome} = require('../controllers/income')
 
-      
-        res.json(totalIncome)
-
-    }   
-    catch (error) {
-        console.log(error );
-        res.json(error)
-    }
-})
+router.get('/total', getIncome)
 
 module.exports = router
